@@ -1,28 +1,31 @@
 {
-  "name": "gulp-rev-all",
-  "version": "0.9.9",
+  "name": "gulp-rev-all2",
+  "version": "3.0.2",
   "description": "Static asset revisioning by appending content hash to filenames: unicorn.css => unicorn.098f6bcd.css, also re-writes references in each file to new reved name. Better absolute path support.",
   "main": "index.js",
   "dependencies": {
-    "chalk": "~0.4.0",
-    "gulp-util": "~2.2.14",
-    "merge": "^1.2.0",
-    "through2": "~0.4.0"
+    "chalk": "^4.1.0",
+    "fancy-log": "^1.3.2",
+    "isbinaryfile": "^4.0.6",
+    "merge": "^2.1.1",
+    "plugin-error": "^1.0.1",
+    "through2": "^4.0.2",
+    "vinyl": "^2.2.1"
   },
   "devDependencies": {
-    "event-stream": "^3.3.4",
-    "gulp": "~3.9.1",
-    "gulp-jshint": "~1.5.1",
-    "gulp-mocha": "~2.2.0",
-    "mocha": "~2.2.1",
-    "should": "*"
+    "eslint": "7.24.0",
+    "event-stream": "^4.0.1",
+    "gulp": "^4.0.2",
+    "mocha": "^8.3.2",
+    "prettier": "^2.2.1",
+    "should": "^13.2.3"
   },
   "scripts": {
-    "test": "mocha test.js"
+    "test": "mocha test.js && eslint ."
   },
   "repository": {
     "type": "git",
-    "url": "git://github.com/malcolmocean/gulp-rev-all.git"
+    "url": "git://github.com/malcolmocean/gulp-rev-all2.git"
   },
   "keywords": [
     "gulp-rev",
@@ -45,13 +48,16 @@
     "cloudfront",
     "cdn"
   ],
-  "homepage": "https://github.com/malcolmocean/gulp-rev-all",
-  "author": "Malcolm Ocean <josh@psidox.com>",
+  "homepage": "https://github.com/malcolmocean/gulp-rev-all2",
+  "author": "Malcolm Ocean <malcolm@malcolmocean.com>",
+  "authors": [
+    "Joshua Bellamy-Henn <josh@psidox.com>"
+  ],
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/malcolmocean/gulp-rev-all/issues"
+    "url": "https://github.com/malcolmocean/gulp-rev-all2/issues"
   },
   "engines": {
-    "node": ">=0.10"
+    "node": ">=10"
   }
 }
