@@ -1,12 +1,12 @@
 requirejs.config({
-    baseUrl: "./",
-    paths: {
-        react: "../lib/react",
-        jquery: "../lib/jquery"
+  baseUrl: "./",
+  paths: {
+    react: "../lib/react",
+    jquery: "../lib/jquery",
+  },
+  shim: {
+    jquery: {
+      exports: "$",
     },
-    "shim": {
-        "jquery" : {
-            "exports": "$"
-        }
-    }
+  },
 });
