@@ -1,7 +1,7 @@
-'use strict';
+"use strict";
 
-exports.no_space_after_map = function() {
-  return 'The purpose of this file is to test regex replace on source map statement'
+exports.no_space_after_map = function () {
+  return "The purpose of this file is to test regex replace on source map statement";
 };
 
 //# sourceMappingURL=no_space_after_map.js.map
