@@ -1,33 +1,32 @@
-var Path = require('path');
-var crypto = require('crypto');
+var Path = require("path");
+var crypto = require("crypto");
+var isBinaryFileSync = require("isbinaryfile").isBinaryFileSync;
 
-module.exports = (function() {
-  'use strict';
+module.exports = (function () {
+  "use strict";
 
-  var path_without_ext = function(path) {
+  var path_without_ext = function (path) {
     var ext = Path.extname(path);
     return path.substr(0, path.length - ext.length);
   };
-  
-  var dirname_with_sep = function(path) {
-    return Path.dirname(path).replace(/\\/g, '/') + '/';
-  }
+
+  var dirname_with_sep = function (path) {
+    return Path.dirname(path).replace(/\\/g, "/") + "/";
+  };
 
   var join_path_url = function (prefix, path) {
-
-    prefix = prefix.replace(/\/$/, '');
-    path = path.replace(/^\//, '');
-    return [ prefix, path ].join('/');
-
+    prefix = prefix.replace(/\/$/, "");
+    path = path.replace(/^\//, "");
+    return [prefix, path].join("/");
   };
 
   /**
    * Joins a directory and a filename, replaces Windows forward-slash with a backslash.
    */
   var join_path = function (directory, filename) {
-
-    return Path.join(directory, filename).replace(/^[a-z]:\\/i, '/').replace(/\\/g, '/');
-    
+    return Path.join(directory, filename)
+      .replace(/^[a-z]:\\/i, "/")
+      .replace(/\\/g, "/");
   };
 
   /**
@@ -35,49 +34,41 @@ module.exports = (function() {
    * Also replaces Windows forward-slash with a backslash.
    */
   var get_relative_path = function (base, path, noStartingSlash) {
-
     if (base === path) {
-      return '';
+      return "";
     }
 
     // Sanitize inputs, convert windows to posix style slashes, ensure trailing slash for base
-    base = base.replace(/^[a-z]:/i, '').replace(/\\/g, '/').replace(/\/$/g, '') + '/';
-    path = path.replace(/^[a-z]:/i, '').replace(/\\/g, '/');
+    base =
+      base
+        .replace(/^[a-z]:/i, "")
+        .replace(/\\/g, "/")
+        .replace(/\/$/g, "") + "/";
+    path = path.replace(/^[a-z]:/i, "").replace(/\\/g, "/");
 
     // Only truncate paths that overap with the base
     if (base === path.substr(0, base.length)) {
-      path = '/' + path.substr(base.length);
+      path = "/" + path.substr(base.length);
     }
 
     var modifyStartingSlash = noStartingSlash !== undefined;
     if (modifyStartingSlash) {
-      if (path[0] === '/' && noStartingSlash) {
+      if (path[0] === "/" && noStartingSlash) {
         path = path.substr(1);
-      } else if (path[0] !== '/' && !noStartingSlash){
-        path = '/' + path;
+      } else if (path[0] !== "/" && !noStartingSlash) {
+        path = "/" + path;
       }
     }
 
     return path;
-
   };
 
-  var md5 = function (str) {
-
-    return crypto.createHash('md5').update(str, 'utf8').digest('hex');
-
+  var md5 = function (buf) {
+    return crypto.createHash("md5").update(buf).digest("hex");
   };
 
   var is_binary_file = function (file) {
-
-    var length = (file.contents.length > 50) ? 50 : file.contents.length;
-    for (var i = 0; i < length; i++) {
-      if (file.contents[i] === 0) {
-        return true;
-      }
-    }
-    return false;
-
+    return isBinaryFileSync(file.contents, file.contents.length);
   };
 
   /**
@@ -85,8 +76,10 @@ module.exports = (function() {
    * the context to the reference file.
    *
    */
-  var get_reference_representations_relative = function (fileCurrentReference, file) {
-
+  var get_reference_representations_relative = function (
+    fileCurrentReference,
+    file
+  ) {
     var representations = [];
 
     //  Scenario 2: Current file is the same directory or lower than the reference
@@ -96,13 +89,29 @@ module.exports = (function() {
     //          file.path = /user/project/second/current_file.html
     //  fileCurrentReference.path = /user/project/second/index.html
 
-    if (dirname_with_sep(fileCurrentReference.path).indexOf(dirname_with_sep(file.path)) === 0) {
-
+    if (
+      dirname_with_sep(fileCurrentReference.path).indexOf(
+        dirname_with_sep(file.path)
+      ) === 0
+    ) {
       //  index.html
-      representations.push(get_relative_path(Path.dirname(file.path), fileCurrentReference.revPathOriginal, true));
+      representations.push(
+        get_relative_path(
+          Path.dirname(file.path),
+          fileCurrentReference.revPathOriginal,
+          true
+        )
+      );
 
       //  ./index.html   (reference: relative)
-      representations.push('.' + get_relative_path(Path.dirname(file.path), fileCurrentReference.revPathOriginal, false));
+      representations.push(
+        "." +
+          get_relative_path(
+            Path.dirname(file.path),
+            fileCurrentReference.revPathOriginal,
+            false
+          )
+      );
     }
 
     //  Scenario 3: Current file is in a different child directory than the reference
@@ -112,20 +121,32 @@ module.exports = (function() {
     //          file.path = /user/project/first/index.html
     //  fileCurrentReference.path = /user/project/second/index.html
 
-    if (dirname_with_sep(file.path) !== dirname_with_sep(fileCurrentReference.path) &&
-      dirname_with_sep(fileCurrentReference.path).indexOf(dirname_with_sep(file.path)) === -1) {
-
-      var pathCurrentReference = dirname_with_sep(get_relative_path(fileCurrentReference.base, fileCurrentReference.revPathOriginal));
-      var pathFile = dirname_with_sep(get_relative_path(file.base, file.revPathOriginal));
+    if (
+      dirname_with_sep(file.path) !==
+        dirname_with_sep(fileCurrentReference.path) &&
+      dirname_with_sep(fileCurrentReference.path).indexOf(
+        dirname_with_sep(file.path)
+      ) === -1
+    ) {
+      var pathCurrentReference = dirname_with_sep(
+        get_relative_path(
+          fileCurrentReference.base,
+          fileCurrentReference.revPathOriginal
+        )
+      );
+      var pathFile = dirname_with_sep(
+        get_relative_path(file.base, file.revPathOriginal)
+      );
 
       // ../second/index.html
       var relPath = Path.relative(pathFile, pathCurrentReference);
-      relPath = relPath.replace(/\\/g, '/');
-      representations.push(relPath + '/' + Path.basename(fileCurrentReference.revPathOriginal));
+      relPath = relPath.replace(/\\/g, "/");
+      representations.push(
+        join_path(relPath, Path.basename(fileCurrentReference.revPathOriginal))
+      );
     }
 
     return representations;
-
   };
 
   /**
@@ -134,18 +155,27 @@ module.exports = (function() {
    *
    */
   var get_reference_representations_absolute = function (fileCurrentReference, file, stripRootPrefixes) {
-
     var representations = [];
     var representation;
 
     //  Scenario 1: Current file is anywhere
     //  /view/index.html  (reference: absolute)
-    representations.push(get_relative_path(fileCurrentReference.base, fileCurrentReference.revPathOriginal, false));
-    
+    representations.push(
+      get_relative_path(
+        fileCurrentReference.base,
+        fileCurrentReference.revPathOriginal,
+        false
+      )
+    );
+
     // Without starting slash, only if it contains a directory
     // view/index.html  (reference: absolute, without slash prefix)
-    representation = get_relative_path(fileCurrentReference.base, fileCurrentReference.revPathOriginal, true);
-    if (representation.indexOf('/')) {
+    representation = get_relative_path(
+      fileCurrentReference.base,
+      fileCurrentReference.revPathOriginal,
+      true
+    );
+    if (representation.indexOf("/")) {
       representations.push(representation);
     }
 
@@ -159,9 +189,7 @@ module.exports = (function() {
     }
 
     return representations;
-
   };
-
 
   return {
     get_relative_path: get_relative_path,
@@ -171,7 +199,6 @@ module.exports = (function() {
     join_path: join_path,
     join_path_url: join_path_url,
     get_reference_representations_relative: get_reference_representations_relative,
-    get_reference_representations_absolute: get_reference_representations_absolute
+    get_reference_representations_absolute: get_reference_representations_absolute,
   };
-
 })();
